Renamed Classes. Connectome is main program now.
